// src/app/models/MortgageInfo.ts
export class MortgageInfo {
    isFirstTimeBuyer: String;
    isItPrimaryResidency: String;
    userId: Number;
    employerName: String;
    jobTitle: String;
    salary: Number;
    addressLine1: String;
    addressLine2: String;
    city: String;
    province: String;
    country: String;
    purchasePrice: Number;
    downPayment: Number;
    mortgageAmount: Number;
    closingDate: Date;
   
}